% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCAdata.R
\docType{data}
\name{brca}
\alias{brca}
\title{TCGA Brest cancer (BRCA) dataset}
\format{
brca is a list of 2 objects:
\enumerate{
\item "seq", a data matrix of 224 rows (subjects) and 17327 columns 
(genes) containing the RNA transcriptomics profiling from TCGA
 pre-processed applying y = log2(x + 0.1) and X = y - med(y);
\item "group", a binary group vector of 112 BRCA patients (1) and 112 
healthy controls (0).
}
}
\source{
\url{https://www.cancer.gov/about-nci/organization/ccg/research/structural-genomics/tcga}
}
\usage{
brca
}
\description{
TCGA data from the "data category": Transcriptome Profiling,
"data type": Gene Expression Quantification, "experimental strategy": RNA
Seq2 and "workflow type": HTSeq-Counts, of 112 brest cancer (BRCA) patients
and 112 healthy controls.
}
\examples{
dim(brca$pc1)
table(brca$group)

}
\references{
Tomczak K, Czerwińska P, Wiznerowicz M. The Cancer Genome Atlas (TCGA):
An Immeasurable Source of Knowledge. Contemporary Oncology 2015; 19 (1A):
A68–A77 <https://10.5114/wo.2014.47136>
}
