% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19.R
\docType{data}
\name{covid19}
\alias{covid19}
\title{COVID-19 RNA-seq dataset}
\format{
covis19 is a data.frame of 15957 rows (genes) and 69 columns (subjects), 
including RSEM count data for 23 (C1-23) non-critical and 46 (P1-46) 
critical COVID-19 cases. The former group is composed by patients in 
a non-critical care ward under supplemental oxygen, while the latter 
included patients in the intensive care unit with Acute Respiratory 
Distress Syndrome (ARDS), as reported by Carapito et al. 2022.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE172114}
}
\usage{
covid19
}
\description{
Expression profiling through high-throughput sequencing 
(RNA-seq, Illumina NovaSeq 6000) of 72 young human patients 
(< 50 years old) with severe forms of COVID-19, 
from Carapito et al. (2022).
}
\examples{
summary(covid19)
dim(covid19)

}
\references{
Carapito R, Li R, Helms J, Carapito C et al. Identification of driver 
genes for critical forms of COVID-19 in a deeply phenotyped young 
patient cohort. Sci Transl Med 2022 Jan 19;14(628):eabj7521.
<https://doi.org/10.1126/scitranslmed.abj7521>
}
