% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTDdata.R
\docType{data}
\name{ftdDNAme}
\alias{ftdDNAme}
\title{Frontotemporal Dementia (FTD) DNA methylation dataset}
\format{
ftdDNAme is a list of 2 objects:
\enumerate{
\item "pc1", a data matrix of 256 rows (subjects) and 16560 columns 
(genes) containing the value of the first principal component of DNAme 
levels, obtained applying a principal component analysis to methylated 
CpG sites within the promoter region, for each gene (genes with 
unmethylated CpGs in both conditions were discarded);
\item "group", a binary group vector of 105 FTD patients (1) and 150 
healthy controls (0).
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE53740}
}
\usage{
ftdDNAme
}
\description{
DNA methylation (DNAme) profiling by genome tiling array 
of 105 FTD patients and 150 healthy controls from peripheral blood 
samples (Li et al., 2014).
}
\examples{
dim(ftdDNAme$pc1)
table(ftdDNAme$group)

}
\references{
Li Y, Chen JA, Sears RL, Gao F et al. An epigenetic signature in 
peripheral blood associated with the haplotype on 17q21.31, a risk 
factor for neurodegenerative tauopathy. PLoS Genet 2014 Mar;10(3):e1004211.
https://doi.org/10.1371/journal.pgen.1004211
}
