% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactomes.R
\docType{data}
\name{ig.DO}
\alias{ig.DO}
\title{DO interactome}
\format{
"ig.DO" is an igraph network object of 6576 nodes and 6923 directed edges, 
corresponding to the Disease Ontology terms (update: Feb, 2020).
It has attributes to vertices/nodes and edges:
\enumerate{
 \item "name" (v/c), i.e., "Term ID";
 \item "term_id" (v/c), i.e., "Term ID";
 \item "term_name" (v/c), i.e,. "Term Name";
 \item "term_distance" (v/c), i.e., Term Distance: the distance to the root; always 0 for the root itself;
 \item "relation" (e/c), either 'is_a' or 'part_of'.
}
}
\source{
\url{https://disease-ontology.org/downloads/}
}
\usage{
ig.DO
}
\description{
DO interactome
}
\examples{

\donttest{
library(SEMgraph)

# DO graph
properties(ig.DO)

}

}
\references{
Schriml et al. Disease Ontology: a backbone for disease semantic integration. Nucleic Acids Res 2012, 40:D940-946.
<https://doi.org/10.1093/nar/gkr972>
}
