% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactomes.R
\docType{data}
\name{ig.GOBP}
\alias{ig.GOBP}
\title{GOBP interactome}
\format{
"ig.GOBP" is an igraph network object of 27685 nodes and 58979 directed edges,
corresponding to the Gene Ontology Biological Process terms (update: Feb, 2020).
It has attributes to vertices/nodes and edges:
\enumerate{
 \item "name" (v/c), i.e., "Term ID";
 \item "term_id" (v/c), i.e., "Term ID";
 \item "term_name" (v/c), i.e,. "Term Name";
 \item "term_distance" (v/c), i.e., Term Distance: the distance to the root; always 0 for the root itself;
 \item "relation" (e/c), either 'is_a' or 'part_of'.
}
}
\source{
\url{https://geneontology.org/docs/download-ontology/}
}
\usage{
ig.GOBP
}
\description{
GOBP interactome
}
\examples{

\donttest{
library(SEMgraph)

# GOBP graph
properties(ig.GOBP)

}

}
\references{
Ashburner et al. Gene ontology: tool for the unification of biology. Nat Genet 2000, 25:25-29.
<https://doi.org/10.1038/75556>
}
