% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactomes.R
\docType{data}
\name{ig.STRING}
\alias{ig.STRING}
\title{STRING interactome}
\format{
"ig.STRING" is an igraph network object of 10829 nodes and 213445 edges 
(undirected), corresponding to the STRING interactome (version 10.5).
Only edges with heigh confidence (combined score>=900) are retained.
}
\source{
\url{https://string-db.org/}
}
\usage{
ig.STRING
}
\description{
STRING interactome
}
\examples{

\donttest{
library(SEMgraph)

# STRING graph
properties(ig.STRING)

}

}
\references{
Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J, 
Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, von Mering C. 
STRING v11: protein-protein association networks with increased 
coverage, supporting functional discovery in genome-wide experimental 
datasets. Nucleic Acids Res. 2019, 47: D607-613. 
<https://doi.org/10.1093/nar/gky1131>
}
