% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipathia.R
\docType{data}
\name{kegg.circuits}
\alias{kegg.circuits}
\title{KEGG circuits}
\format{
"kegg.circuits" is a list of igraph objects corresponding to the "hsa###" KEGG
signalling pathways. This list includes 146 pathways, which are subdivided into "effector"
subpathays (i.e., a list  1868 subgraphs with many different input nodes and one output node),
and "decomposed" subpathays(i.e., a list 8409 subgraphs with one input node and one output node).
}
\source{
\url{https://bioconductor.org/packages/hipathia/}
}
\usage{
kegg.circuits
}
\description{
KEGG circuits extracted using the \code{hipathia} R package
(update: October, 2024).
}
\examples{

\donttest{
library(SEMgraph)

str(kegg.circuits, max.level=1) #list of 6
#str(kegg.circuits$pathigraphs, max.level=1) #list of 145
str(kegg.circuits$pathigraphs[[2]], max.level=1) #list of 11

start <- Sys.time()
for (k in 2:2) { #k in 1:length(kegg.circuits$pathigraphs)
 name <- kegg.circuits$pathigraphs[[k]]$path.name
 print(name)
 
 # (a)-signalling pathway
 gk <- kegg.circuits$pathigraphs[[k]]$graph
 gplot(gk, main = name)
 Sys.sleep(1)
 
   # (b)-"effector" subpathways
   for (j in 1:length(kegg.circuits$pathigraphs[[k]]$effector.subgraphs)) {
    cat(" ", j, "\n")
    sgj <- kegg.circuits$pathigraphs[[j]]$effector.subgraphs[[j]]
    gplot(sgj)
    Sys.sleep(1)
   }
	
    # (c)-"decomposed" subpathways
    for (i in 1:length(kegg.circuits$pathigraphs[[k]]$subgraphs)) {
     cat("   ", i, "\n")
     sgi <- kegg.circuits$pathigraphs[[k]]$subgraphs[[i]]
     gplot(sgi)
     Sys.sleep(1)
    }
    cat("\n")
}
end <- Sys.time()
print(end-start)
 
}

}
\references{
Kanehisa M, Goto S (1999). KEGG: kyoto encyclopedia of genes and genomes. 
Nucleic Acid Research 28(1): 27-30. 
<https://doi.org/10.1093/nar/27.1.29>

Hidalgo MR, Cubuk C, Amadoz A, Salavert F, Carbonell-Caballero J, Dopazo J. High
throughput estimation of functional cell activities reveals disease mechanisms
and predicts relevant clinical outcomes. Oncotarget. 2017 Jan 17;8(3):5160-5178.
<https://doi:10.18632/oncotarget.14107>
}
