% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\docType{data}
\name{kegg.pathways}
\alias{kegg.pathways}
\title{KEGG signaling pathways collection}
\format{
"kegg.pathways" is a list of 306 KEGG signaling pathways, stored as 
igraph objects.
}
\source{
\url{https://www.kegg.jp/kegg}
}
\usage{
kegg.pathways
}
\description{
Collection of KEGG signaling pathways extracted using 
the \code{graphite} R package (update: April, 2020).
}
\examples{

## NOT RUN ## {

# Number of nodes per pathway
kegg.nodes <- unlist(lapply(kegg.pathways, vcount))
names(kegg.nodes) <- NULL
# Number of edges per pathway
kegg.edges <- unlist(lapply(kegg.pathways, ecount))
names(kegg.edges) <- NULL
# Gene list per pathway
kegg.genes <- unlist(lapply(kegg.pathways, function(x) V(x)$name))
quantile(kegg.nodes)
quantile(kegg.edges)
length(unique(kegg.genes))  # Number of unique genes within the dataset

# Loading breast cancer KEGG network
i <- which(names(kegg.pathways) == "Steroid biosynthesis")
sb.graph <- kegg.pathways[[i]]
summary(sb.graph)
gplot(sb.graph)

## }

}
\references{
Kanehisa M, Goto S (1999). KEGG: kyoto encyclopedia of genes and genomes. 
Nucleic Acid Research 28(1): 27-30. 
https://doi.org/10.1093/nar/27.1.29

Martini P, Sales G, Massa MS, Chiogna M, Romualdi C. Along signal paths: 
anempirical gene set approach exploiting pathway topology. 
Nucleic Acids Res. 41(1):e19.
https://doi.org/10.1093/nar/gks866.
}
