% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomePathways.R
\docType{data}
\name{reactome.pathways}
\alias{reactome.pathways}
\title{Reactome pathways collection}
\format{
"reactome.pathways" is a list of 1641 Reactome pathways, stored as 
igraph objects.
}
\source{
\url{https://reactome.org/}
}
\usage{
reactome.pathways
}
\description{
Collection of Reactome signaling pathways extracted using 
the \code{graphite} R package (update: April, 2020).
}
\examples{

\donttest{
library(SEMgraph)

# Reactome pathways
names(reactome.pathways)

# Loading NOTCH4 graph
i <- which(names(reactome.pathways) == "NOTCH4 Intracellular Domain Regulates Transcription")
ig <- reactome.pathways[[i]]
properties(ig)
gplot(ig, l="dot", psize=50, main=names(reactome.pathways[i]))

}

}
\references{
Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, 
Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, 
Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, 
Wu G, Stein L, Hermjakob H, D'Eustachio P. 
The reactome pathway knowledgebase. Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. 
<https://doi.org/10.1093/nar/gkz1031>

Sales G, Calura E, Cavalieri D, Romualdi C (2012). graphite - a Bioconductor
package to convert pathway topology to gene network. BMC Bioinformatics 2012,13:20.
<https://doi.org/10.1186/1471-2105-13-20>
}
