% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactomes.R
\docType{data}
\name{reactome}
\alias{reactome}
\title{Reactome interactome}
\format{
"reactome" is an igraph network object of 9762 nodes and 416128 edges 
(109268 directed and 306860/2 = 153430 bidirected) corresponding to the union
of 1641 Reactome pathways.
}
\source{
\url{https://reactome.org/}
}
\usage{
reactome
}
\description{
Interactome generated by merging Reactome pathways 
extracted using the \code{graphite} R package (update: April, 2020).
}
\examples{

\donttest{
library(SEMgraph)

#Reactome graph
properties(reactome)

}

}
\references{
Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, 
Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, 
Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, 
Wu G, Stein L, Hermjakob H, D'Eustachio P. The reactome pathway knowledgebase. 
Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503.
<https://doi:10.1093/nar/gkz1031>.

Sales G, Calura E, Cavalieri D, Romualdi C (2012). graphite - a Bioconductor
package to convert pathway topology to gene network. BMC Bioinformatics 2012, 13:20.
<https://doi:10.1186/1471-2105-13-20>.
}
