% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomeInteractome.R
\docType{data}
\name{reactome}
\alias{reactome}
\title{Reactome interactome}
\format{
"reactome" is an igraph network object of 9762 nodes and 416128 edges 
corresponding to the union of 1641 Reactome pathways.
}
\source{
\url{https://reactome.org}
}
\usage{
reactome
}
\description{
Interactome generated by merging Reactome pathways 
extracted using the \code{graphite} R package (update: April, 2020).
}
\examples{

summary(reactome)

# Reactome degrees of freedom
vcount(reactome)*(vcount(reactome) - 1)/2 - ecount(reactome)

# Reactome average shortest path length
mean_distance(reactome)
}
\references{
Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, 
Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, 
Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, 
Wu G, Stein L, Hermjakob H, D'Eustachio P. 
The reactome pathway knowledgebase. 
Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. PubMed PMID: 31691815.
<doi:10.1093/nar/gkz1031>

Martini P, Sales G, Massa MS, Chiogna M, Romualdi C. Along signal paths: 
anempirical gene set approach exploiting pathway topology. 
Nucleic Acids Res. 41(1):e19.
<https://doi.org/10.1093/nar/gks866>
}
